Fix: Farm add issue

See merge request courses/2024-winter/csci5308/Group09!22

+++ backend/src/main/java/com/example/backend/entities/Farms.java
@@ -19,8 +19,8 @@ public class Farms {
 
     private String name;
     private String Address;
-    private int lat;
-    private int lng;
+    private double lat;
+    private double lng;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
